+++ queue_job/__init__.py
@@ -1,5 +1,6 @@
 from . import controllers
 from . import fields
 from . import models
+from . import wizards
 from . import jobrunner
 from .hooks.post_init_hook import post_init_hook
