+++ queue_job/jobrunner/__init__.py
@@ -50,7 +50,7 @@ class WorkerJobRunner(server.Worker):
     """ Jobrunner workers """
 
     def __init__(self, multi):
-        super(WorkerJobRunner, self).__init__(multi)
+        super().__init__(multi)
         self.watchdog_timeout = None
         self.runner = QueueJobRunner.from_environ_or_config()
 
@@ -59,7 +59,7 @@ class WorkerJobRunner(server.Worker):
 
     def signal_handler(self, sig, frame):
         _logger.debug("WorkerJobRunner (%s) received signal %s", self.pid, sig)
-        super(WorkerJobRunner, self).signal_handler(sig, frame)
+        super().signal_handler(sig, frame)
         self.runner.stop()
 
     def process_work(self):
