+++ queue_job/tests/common.py
@@ -289,7 +289,7 @@ class JobsTrap():
 class JobCounter:
 
     def __init__(self, env):
-        super(JobCounter, self).__init__()
+        super().__init__()
         self.env = env
         self.existing = self.search_all()
